Clarify "header notice" is just "brexit notice"
Previously we had a generic "header notice" component that was only
actually used once. While we should aim for generic components, the
behaviour of this one is clearly coupled to the needs of Brexit. By
clearly marking it as a "brexit thing", we:

- Make it easier to identify for removal, along with other bits of
code that were only added around the transition period.

- Make it more flexible, without having to worry about accomodating
other use cases (since there are none).
