+++ app/views/components/_brexit-notice.html.erb
@@ -1,0 +1,46 @@
+<%
+  title ||= false
+  description ||= false
+  link_intro ||= false
+  links ||= []
+  featured_link ||= false
+%>
+<% if title && description %>
+  <section class="app-c-brexit-notice" aria-label="notice" role="region" lang="en">
+    <div class="app-c-brexit-notice__header">
+      <h2 class="app-c-brexit-notice__title"><%= title %></h2>
+    </div>
+
+    <div class="app-c-brexit-notice__content">
+      <% if description.is_a? Array %>
+        <% description.map do |line| %>
+          <p class="app-c-brexit-notice__text govuk-body"><%= line %></p>
+        <% end %>
+      <% else %>
+        <p class="app-c-brexit-notice__text govuk-body"><%= description %></p>
+      <% end %>
+      <% if featured_link %>
+        <p class="app-c-brexit-notice__text govuk-body"><%= featured_link %></p>
+      <% end %>
+      <% if links.any? %>
+        <div class="app-c-brexit-notice__links">
+          <% if links.count == 1 %>
+            <p class="govuk-body app-c-brexit-notice__link-intro">
+              <% if link_intro %><%= link_intro %><% end %>
+              <%= link_to(links[0][:title], links[0][:href], data: links[0][:data_attributes], class: "app-c-brexit-notice__link") %>
+            </p>
+          <% else %>
+            <% if link_intro %><p class="govuk-body app-c-brexit-notice__link-intro"><%= link_intro %></p><% end %>
+            <ul class="app-c-brexit-notice__list govuk-list">
+              <% links.each do |link| %>
+                <li><%= link_to(link[:title], link[:href], data: link[:data_attributes], class: "app-c-brexit-notice__link") %></li>
+              <% end %>
+            </ul>
+          <% end %>
+        </div>
+      <% end %>
+
+
+    </div>
+  </section>
+<% end %>
