+++ app/views/components/docs/brexit-notice.yml
@@ -1,4 +1,4 @@
-name: Header Notice
+name: Brexit Notice
 description: A bolder style of notice
 shared_accessibility_criteria:
   - link
