+++ app/views/content_items/case_study.html.erb
@@ -16,7 +16,7 @@
 
 <%= render 'shared/publisher_metadata_with_logo' %>
 <%= render 'govuk_publishing_components/components/notice', @content_item.withdrawal_notice_component  %>
-<%= render 'components/header-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
+<%= render 'components/brexit-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds content-bottom-margin">
