+++ app/views/content_items/document_collection.html.erb
@@ -15,7 +15,7 @@
   <div class="govuk-grid-column-two-thirds">
     <%= render 'govuk_publishing_components/components/lead_paragraph', text: @content_item.description %>
     <%= render 'govuk_publishing_components/components/notice', @content_item.withdrawal_notice_component  %>
-    <%= render 'components/header-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
+    <%= render 'components/brexit-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
     <%= render 'shared/history_notice', content_item: @content_item %>
   </div>
 </div>
