+++ app/views/content_items/html_publication.html.erb
@@ -33,7 +33,7 @@
 <% end %>
 
 <%= render 'govuk_publishing_components/components/notice', @content_item.withdrawal_notice_component  %>
-<%= render 'components/header-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
+<%= render 'components/brexit-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
 
 <div
   class="govuk-grid-row sidebar-with-body"
