+++ app/views/content_items/publication.html.erb
@@ -23,7 +23,7 @@
 <%= render 'shared/history_notice', content_item: @content_item %>
 
 <%= render 'govuk_publishing_components/components/notice', @content_item.withdrawal_notice_component  %>
-<%= render 'components/header-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
+<%= render 'components/brexit-notice', @content_item.brexit_notice_component unless @content_item.withdrawal_notice_component %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds responsive-bottom-margin">
